+++ src/EditorFacade.js
@@ -1,8 +1,6 @@
 import * as languagesJson from './configuration/languages.json';
-import * as languagesJsonV3 from './configuration/languagesV3.json';
 import { editorLogger as logger } from './configuration/LoggerConfig';
 import { Editor } from './Editor';
-import * as DefaultConfiguration from './configuration/DefaultConfiguration';
 
 /**
  * Attach an Editor to a DOMElement
@@ -24,6 +22,5 @@ export function register(element, configuration, penStyle, theme, behaviors) {
  * @return {JSON} A list of available languages
  */
 export function getAvailableLanguageList(configuration) {
-  const innerConfiguration = DefaultConfiguration.overrideDefaultConfiguration(configuration);
-  return innerConfiguration.recognitionParams.apiVersion === 'V4' ? languagesJson : languagesJsonV3;
+  return languagesJson;
 }